fix: rename types to telemetry to avoid shadowing
https://github.com/dagster-io/dagster/issues/21920
